requirements.txt
Flask==3.1.1
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
python-dotenv==1.1.1
Requests==2.32.4
user_agents==2.2.0
Werkzeug==3.1.3
